fale, how works works;

+++ routes/userRoutes.js
@@ -51,8 +51,7 @@ router.post('/api/newuser', async (req, res) => {
       to: req.body.Email,
       from: 'mntasks@interia.pl',
       subject: 'nmTasks Confrim your account',
-      text: 'test',
-      html: `<div style="font-family:Verdana;margin: 0 auto; padding: 0; max-width: 90%; align-items: center; justify-content: center; flex: 1 1 100%; flex-flow: column wrap; border-radius: 10px; color: #353353; font-size: 10px; "> <h3 style=" max-width: 100%; background-color: #1db95e; color: #fff; font-size: 3em; padding: 1em 0 1em 1em; border-radius: 10px 10px 0 0; margin: 0; "> mnTasks </h3> <div style="text-align: left; color: #353353;font-size: 1.1em; background-color: #f9f9f9"> <h5 style=" font-size: 2em; font-weight: 700; padding: 1em 0; margin: 0 0 0 0.5em;color: #353353; "> Reset your password </h5> <div style="margin: 0 0 1em 1em"> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 0"> We recently received a request to reset the password to your account: </p> <span style="font-weight: 700;font-size:1.3em">${req.body.Email}</span> </div> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 1em"> To reset your password, click on the button bellow or copy and paste link into your browse </p> <p style="font-size:1.3em;margin: 0 0 1em 1em; color:#4285F4;text-decoration: underline">https://mntasks.herokuapp.com/reset/${t}</p> <div> <a href="https://mntasks.herokuapp.com/reset/${t}" target="_blank" style="display:inline-block; padding: 0.75em 1.2em; margin: 0.5em 0 2.5em 1em; font-size: 1.2em; font-weight: 700; background: #1db95e;text-decoration:none; background-color:#1db95e;border: none; border-radius: 0.3em; color: #fff; "> Reset </a> </div> </div> <div style="text-align: left; background-color: #E6E6E6; padding: 0; border-radius: 0 0 10px 10px;"> <p style="font-weight: bold; color: #353353;font-size: 1.4em; margin:0 0 0 1em; padding-top: 1em;"> Important Security Notice: </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> mnTaska never asks for your password or other sensitive information by email. Do not click links or respond to a suspicious email! </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> Do not reply to this email! This message was generated automaticly. </p> <p style="margin-left: 1em;color: #353353;padding-bottom: 1em; font-size: 1.3em;"> &copy; mnTasks 2021 </p> </div> </div> `,
+      html: `<div style="font-family:Verdana;margin: 0 auto; padding: 0; max-width: 90%; align-items: center; justify-content: center; flex: 1 1 100%; flex-flow: column wrap; border-radius: 10px; color: #353353; font-size: 10px; "> <h3 style=" max-width: 100%; background-color: #1db95e; color: #fff; font-size: 3em; padding: 1em 0 1em 1em; border-radius: 10px 10px 0 0; margin: 0; "> mnTasks </h3> <div style="text-align: left; color: #353353;font-size: 1.1em; background-color: #f9f9f9"> <h5 style=" font-size: 2em; font-weight: 700; padding: 1em 0; margin: 0 0 0 0.5em;color: #353353; "> Confirm your account </h5> <div style="margin: 0 0 1em 1em"> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 0"> To confirm your account <span style="font-weight: 700;font-size:1em">${req.body.Email}</span> please clik the link below or copy and paste link into your browse </p> </div> <p style="font-size:1.3em;margin: 0 0 1em 1em; color:#4285F4;text-decoration: underline">https://mntasks.herokuapp.com/almost-there/${t}</p> <div><a href="href='https://mntasks.herokuapp.com/almost-there/${t}'" target="_blank" style="display:inline-block; padding: 0.75em 1.2em; margin: 0.5em 0 2.5em 1em; font-size: 1.2em; font-weight: 700; background: #1db95e;text-decoration:none; background-color:#1db95e;border: none; border-radius: 0.3em; color: #fff; "> Confirm now </a> </div> </div> <div style="text-align: left; background-color: #E6E6E6; padding: 0; border-radius: 0 0 10px 10px;"> <p style="font-weight: bold; color: #353353;font-size: 1.4em; margin:0 0 0 1em; padding-top: 1em;"> Important Security Notice: </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> mnTaska never asks for your password or other sensitive information by email. Do not click links or respond to a suspicious email! </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> Do not reply to this email! This message was generated automaticly. </p> <p style="margin-left: 1em;color: #353353;padding-bottom: 1em; font-size: 1.3em;"> &copy; mnTasks 2021 </p> </div> </div>`,
     };
 
     sgMail.send(msg);
@@ -96,7 +95,7 @@ router.post('/api/resetPassword', (req, res) => {
           from: 'mntasks@interia.pl',
           subject: 'nmTasks Reset your password',
           text: 'We have sent mail to reset password',
-          html: `<div style="font-family:Verdana;margin: 0 auto; padding: 0; max-width: 90%; align-items: center; justify-content: center; flex: 1 1 100%; flex-flow: column wrap; border-radius: 10px; color: #353353; font-size: 10px; "> <h3 style=" max-width: 100%; background-color: #1db95e; color: #fff; font-size: 3em; padding: 1em 0 1em 1em; border-radius: 10px 10px 0 0; margin: 0; "> mnTasks </h3> <div style="text-align: left; color: #353353;font-size: 1.1em; background-color: #f9f9f9"> <h5 style=" font-size: 2em; font-weight: 700; padding: 1em 0; margin: 0 0 0 0.5em;color: #353353; "> Reset your password </h5> <div style="margin: 0 0 1em 1em"> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 0"> We recently received a request to reset the password to your account: </p> <span style="font-weight: 700;font-size:1.3em">${req.body.Email}</span> </div> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 1em"> To reset your password, click on the button bellow or copy and paste link into your browse </p> <p style="font-size:1.3em;margin: 0 0 1em 1em; color:#4285F4;text-decoration: underline">https://mntasks.herokuapp.com/reset/${t}</p> <div> <a href="https://mntasks.herokuapp.com/reset/${t}" target="_blank" style="display:inline-block; padding: 0.75em 1.2em; margin: 0.5em 0 2.5em 1em; font-size: 1.2em; font-weight: 700; background: #1db95e;text-decoration:none; background-color:#1db95e;border: none; border-radius: 0.3em; color: #fff; "> Reset </a> </div> </div> <div style="text-align: left; background-color: #E6E6E6; padding: 0; border-radius: 0 0 10px 10px;"> <p style="font-weight: bold; color: #353353;font-size: 1.4em; margin:0 0 0 1em; padding-top: 1em;"> Important Security Notice: </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> mnTaska never asks for your password or other sensitive information by email. Do not click links or respond to a suspicious email! </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> Do not reply to this email! This message was generated automaticly. </p> <p style="margin-left: 1em;color: #353353;padding-bottom: 1em; font-size: 1.3em;"> &copy; mnTasks 2021 </p> </div> </div> `,
+          html: `<div style="font-family:Verdana;margin: 0 auto; padding: 0; max-width: 90%; align-items: center; justify-content: center; flex: 1 1 100%; flex-flow: column wrap; border-radius: 10px; color: #353353; font-size: 10px; "> <h3 style=" max-width: 100%; background-color: #1db95e; color: #fff; font-size: 3em; padding: 1em 0 1em 1em; border-radius: 10px 10px 0 0; margin: 0; "> mnTasks </h3> <div style="text-align: left; color: #353353;font-size: 1.1em; background-color: #f9f9f9"> <h5 style=" font-size: 2em; font-weight: 700; padding: 1em 0; margin: 0 0 0 0.5em;color: #353353; "> Reset your password </h5> <div style="margin: 0 0 1em 1em"> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 0"> We recently received a request to reset the password to your account: </p> <span style="font-weight: 700;font-size:1.3em">${req.body.Email}</span> </div> <p style="font-size:1.3em;font-weight: 500;color: #353353; margin: 0 0 1em 1em"> To reset your password, click on the button bellow or copy and paste link into your browse </p> <p style="font-size:1.3em;margin: 0 0 1em 1em; color:#4285F4;text-decoration: underline">https://mntasks.herokuapp.com/reset/${t}</p> <div> <a href="https://mntasks.herokuapp.com/reset/${t}" target="_blank" style="display:inline-block; padding: 0.75em 1.2em; margin: 0.5em 0 2.5em 1em; font-size: 1.2em; font-weight: 700; background: #1db95e;text-decoration:none; background-color:#1db95e;border: none; border-radius: 0.3em; color: #fff; "> Reset </a> </div> </div> <div style="text-align: left; background-color: #E6E6E6; padding: 0; border-radius: 0 0 10px 10px;"> <p style="font-weight: bold; color: #353353;font-size: 1.4em; margin:0 0 0 1em; padding-top: 1em;"> Important Security Notice: </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> mnTaska never asks for your password or other sensitive information by email. Do not click links or respond to a suspicious email! </p> <p style="margin-left: 1em;color: #353353; font-size: 1.3em;"> Do not reply to this email! This message was generated automaticly. </p> <p style="margin-left: 1em;color: #353353;padding-bottom: 1em; font-size: 1.3em;"> &copy; mnTasks 2021 </p> </div> </div>`,
         };
 
         sgMail.send(msg);
@@ -252,6 +251,13 @@ router.get('/api/todos', async (req, res) => {
 });
 
 router.post('/api/delete', async (req, res) => {
+  const client = await Users.findOne({ _id: req.session.user._id });
+
+  const { password } = req.body;
+
+  const validPass = await bcrypt.compare(password, client.password);
+  if (!validPass) return res.send({ message: 'Invalid password!' });
+
   const userToDelete = await Users.deleteOne({ _id: req.session.user._id });
   const deleteUserTasks = await Todo.deleteMany({
     userId: req.session.user._id,
@@ -259,7 +265,7 @@ router.post('/api/delete', async (req, res) => {
 
   delete req.session.user;
   res.clearCookie('user');
-  res.send('logged out');
+  res.send({ message: 'Your account has been deleted', correct: true });
 });
 
 router.get('/api/userpanel', async (req, res) => {
@@ -336,7 +342,7 @@ router.post('/api/editData', async (req, res) => {
         return res.send({ message: 'Old nickname cannot be a new nickname' });
 
       user.name = data;
-      res.send({message: 'You successful change nickname',correct: true})
+      res.send({ message: 'You successful change nickname', correct: true });
       break;
     case 'email':
       const { error } = editDataValidation(req.body);
@@ -350,7 +356,7 @@ router.post('/api/editData', async (req, res) => {
       if (userEmail) return res.send({ message: 'Email adress exists!' });
 
       user.email = data;
-      res.send({message: 'You successful change email',correct: true})
+      res.send({ message: 'You successful change email', correct: true });
       break;
     case 'password':
       if (data.length < 6)
@@ -363,7 +369,7 @@ router.post('/api/editData', async (req, res) => {
       const salt = await bcrypt.genSalt(10);
       const hashPassword = await bcrypt.hash(data, salt);
       user.password = hashPassword;
-      res.send({message: 'You successful change password',correct: true})
+      res.send({ message: 'You successful change password', correct: true });
       break;
   }
   user.save();
